+++ src/lib/accounts/Set-CosmosDbAccount.ps1
@@ -46,6 +46,10 @@ function Set-CosmosDbAccount
         $IpRangeFilter,
 
         [Parameter()]
+        [System.String[]]
+        $AllowedOrigin,
+
+        [Parameter()]
         [Switch]
         $AsJob
     )
@@ -58,6 +62,7 @@ function Set-CosmosDbAccount
     $null = $getCosmosDbAccount_parameters.Remove('MaxIntervalInSeconds')
     $null = $getCosmosDbAccount_parameters.Remove('MaxStalenessPrefix')
     $null = $getCosmosDbAccount_parameters.Remove('IpRangeFilter')
+    $null = $getCosmosDbAccount_parameters.Remove('AllowedOrigin')
     $null = $getCosmosDbAccount_parameters.Remove('AsJob')
     $existingAccount = Get-CosmosDbAccount @getCosmosDbAccount_parameters
 
@@ -132,12 +137,35 @@ function Set-CosmosDbAccount
         ipRangeFilter            = $ipRangeFilterString
     }
 
+    if ($PSBoundParameters.ContainsKey('AllowedOrigin'))
+    {
+        $allowedOriginString = ($AllowedOrigin -join ',')
+    }
+    else
+    {
+        $allowedOriginString = $existingAccount.Properties.cors.allowedOrigins
+    }
+
+    if (-not ([System.String]::IsNullOrEmpty($allowedOriginString)))
+    {
+        $corsObject = @(
+            @{
+                allowedOrigins = $allowedOriginString
+            }
+        )
+
+        $cosmosDBProperties += @{
+            cors = $corsObject
+        }
+    }
+
     $null = $PSBoundParameters.Remove('Location')
     $null = $PSBoundParameters.Remove('LocationRead')
     $null = $PSBoundParameters.Remove('DefaultConsistencyLevel')
     $null = $PSBoundParameters.Remove('MaxIntervalInSeconds')
     $null = $PSBoundParameters.Remove('MaxStalenessPrefix')
     $null = $PSBoundParameters.Remove('IpRangeFilter')
+    $null = $PSBoundParameters.Remove('AllowedOrigin')
 
     $setAzResource_parameters = $PSBoundParameters + @{
         ResourceType = 'Microsoft.DocumentDb/databaseAccounts'